Make Java dependency optional (#7)
* Change Java dependency to optional

* Fix bug with result reasons

+++ src/main/kotlin/com/codacy/intellij/plugin/services/api/models/GetRepositoryPullRequestResponse.kt
@@ -52,7 +52,7 @@ data class DiffCoverage(
 
 data class ResultReason(
     val gate: String,
-    val expected: Int,
+    val expected: Double,
     val expectedThreshold: ExpectedThreshold,
     val isUpToStandards: Boolean
 )
